% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_min.R
\name{theme_powerpoint}
\alias{theme_powerpoint}
\title{theme_powerpoint}
\usage{
theme_powerpoint(base_size = 11, base_family = "")
}
\description{
A clean theme for ggplot2, with some handy defaults for Powerpoints:
essentially \code{\link{theme_min}} with larger text in certain elements,
and a little room between axis titles and the plot.
}
